+++ osu.Game.Rulesets.Taiko/Edit/Blueprints/DrumRollPlacementBlueprint.cs
@@ -1,0 +1,12 @@
+using osu.Game.Rulesets.Taiko.Objects;
+
+namespace osu.Game.Rulesets.Taiko.Edit.Blueprints
+{
+    public class DrumRollPlacementBlueprint : TaikoSpanPlacementBlueprint
+    {
+        public DrumRollPlacementBlueprint()
+            : base(new DrumRoll())
+        {
+        }
+    }
+}
