+++ osu.Game.Rulesets.Taiko/Edit/Blueprints/SwellPlacementBlueprint.cs
@@ -1,0 +1,12 @@
+using osu.Game.Rulesets.Taiko.Objects;
+
+namespace osu.Game.Rulesets.Taiko.Edit.Blueprints
+{
+    public class SwellPlacementBlueprint : TaikoSpanPlacementBlueprint
+    {
+        public SwellPlacementBlueprint()
+            : base(new Swell())
+        {
+        }
+    }
+}
